Add Codex module with ApplyPatch functionality
- Introduced a new Codex module with a Cargo.toml and CMakeLists.txt for Rust integration.
- Implemented ApplyPatchShard to apply OpenAI Codex patches to files.
- Removed ApplyPatchShard implementation from the Markdown module to streamline functionality.
- Updated dependencies in the Markdown module's Cargo.toml to reflect changes.
