+++ shards/modules/codex/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "shards-markdown"
+name = "shards-codex"
 description = ""
 license = "BSD-3-Clause"
 version = "0.1.0"
@@ -13,5 +13,4 @@ crate-type = ["rlib", "staticlib"]
 lazy_static = "1.5"
 shards = { path = "../../rust" }
 compile-time-crc32 = "0.1"
-pulldown-cmark = { version = "0.13", default-features = false }
-htmd = "0.2.1"
+codex-apply-patch = { path = "../../../deps/codex/codex-rs/apply-patch" }
